Basic peer discovery (p2p)
- Send listening port to remote peers
- Handle listening port in mock backend, send Discovered event from mock backend
- Allow only one connection per socket address
- Return inbound flag from RPC in get_connected_peers
- Discover remote addresses for incoming connections only.
- Remove excessive trait bounds from tests
- Allow specifying initial nodes to connect
- Allow binding to multiple addresses and return all listening addresses from RPC
- Use socket2 in mock backend
- Add Ip4 and Ip6 types
- Make PeerDb to not depend on globally unique ids
- Store connected peers and available addresses separately in PeerDb.

+++ p2p/src/testing_utils.rs
@@ -62,7 +62,7 @@ pub struct TestTransportChannel {}
 impl TestTransportMaker for TestTransportChannel {
     type Transport = MockChannelTransport;
 
-    type Address = u64;
+    type Address = u32;
 
     fn make_transport() -> Self::Transport {
         MockChannelTransport::new()
@@ -124,7 +124,7 @@ impl RandomAddressMaker for TestTcpAddressMaker {
 pub struct TestChannelAddressMaker {}
 
 impl RandomAddressMaker for TestChannelAddressMaker {
-    type Address = u64;
+    type Address = u32;
 
     fn new() -> Self::Address {
         let mut rng = make_pseudo_rng();
